rework plotter.py
