+++ python/src/robyn/visualization/plotter.py
@@ -1,0 +1,115 @@
+from typing import List, Dict, Any, Optional
+import pandas as pd
+import matplotlib.pyplot as plt
+
+from robyn.data.entities.mmmdata import MMMData
+from robyn.data.entities.hyperparameters import Hyperparameters, ChannelHyperparameters
+from robyn.modeling.feature_engineering import FeaturizedMMMData  # New import
+
+
+class FeaturePlotter:
+    """
+    A class for creating various plots related to feature engineering in the Robyn framework.
+    """
+
+    def __init__(self, mmm_data: MMMData, hyperparameters: Hyperparameters):
+        """
+        Initialize the FeaturePlotter class.
+
+        Args:
+            mmm_data (MMMData): The input data and specifications for the MMM.
+            hyperparameters (Hyperparameters): The hyperparameters for the model.
+        """
+        self.mmm_data = mmm_data
+        self.hyperparameters = hyperparameters
+
+    def plot_adstock(self, channel: str) -> plt.Figure:
+        """
+        Plot the adstock transformation for a specific channel.
+
+        Args:
+            channel (str): The name of the channel to plot adstock for.
+
+        Returns:
+            plt.Figure: A matplotlib Figure object containing the adstock plot.
+        """
+        pass
+
+    def plot_saturation(self, channel: str) -> plt.Figure:
+        """
+        Plot the saturation curve transformation for a specific channel.
+
+        Args:
+            channel (str): The name of the channel to plot saturation for.
+
+        Returns:
+            plt.Figure: A matplotlib Figure object containing the saturation curves plot.
+        """
+        pass
+
+    def plot_spend_exposure(self, featurized_data: FeaturizedMMMData, channel: str) -> plt.Figure:
+        """
+        Plot the relationship between spend and exposure for a given channel.
+
+        Args:
+            featurized_data (FeaturizedMMMData): The featurized data after feature engineering.
+            channel (str): The name of the channel being plotted.
+
+        Returns:
+            plt.Figure: A matplotlib Figure object containing the spend-exposure plot.
+        """
+        dt_mod = featurized_data.dt_mod
+        # Rest of the method implementation
+        pass
+
+    def plot_feature_importance(self, feature_importance: Dict[str, float]) -> plt.Figure:
+        """
+        Plot the importance of different features in the model.
+
+        Args:
+            feature_importance (Dict[str, float]): Dictionary of feature importances.
+
+        Returns:
+            plt.Figure: A matplotlib Figure object containing the feature importance plot.
+        """
+        pass
+
+    def plot_response_curves(self, featurized_data: FeaturizedMMMData) -> Dict[str, plt.Figure]:
+        """
+        Plot response curves for different channels.
+
+        Args:
+            featurized_data (FeaturizedMMMData): The featurized data after feature engineering.
+
+        Returns:
+            Dict[str, plt.Figure]: Dictionary mapping channel names to their respective response curve plots.
+        """
+        dt_mod = featurized_data.dt_mod
+        # Rest of the method implementation
+        pass
+
+    def save_plots(self, plots: Dict[str, plt.Figure], output_dir: str) -> None:
+        """
+        Save all generated plots to the specified output directory.
+
+        Args:
+            plots (Dict[str, plt.Figure]): Dictionary mapping plot names to matplotlib Figure objects.
+            output_dir (str): Directory path where the plots should be saved.
+
+        Returns:
+            None
+        """
+        pass
+
+    def _apply_transformations(self, x: pd.Series, params: ChannelHyperparameters) -> pd.Series:
+        """
+        Apply adstock and saturation transformations to the input series.
+
+        Args:
+            x (pd.Series): Input series to transform.
+            params (ChannelHyperparameters): Hyperparameters for the channel.
+
+        Returns:
+            pd.Series: Transformed series.
+        """
+        pass
